Update README.md
In-state vs Out-of-state Bill

The University of South by Southwest charges $3000 per semester for in-state tuition and $4500 per semester for out-of-state tuition. In addition, room and board is $2500 per semester for in-state students and $3500 per semester for out-of-state students.

Write a program that first prompts the user for their residential status. You can use a prompt like this and store their answer as an integer.

Enter 1 for in-state residency or 2 for out-of-state:

Then ask them whether they require room and board. You can use a prompt like this and store their answer as an integer.

Enter 1 if you need room and board or 2 for no room and board:

The program should then compute and output their bill for that semester. A sample program has been started for you that shows how the && operator can be used in an if statement.

Here are two sample runs of the program for two different inputs:

Sample Run 1:

Enter 1 for in-state residency or 2 for out-of-state: 1
Enter 1 for if you need room and board or 2 for no room and board: 2
Your total bill for this semester is $3000

Sample Run 2:

Enter 1 for in-state residency or 2 for out-of-state: 2
Enter 1 for if you need room and board or 2 for no room and board: 1
Your total bill for this semester is $8000